chore: update demo banner image ```
Since this is a binary file change (an image update) that doesn't fit into any of the other specific categories, I've used "chore" as the type. The description is kept simple and concise since the change is straightforward - updating an image file. There's no need for a body in this case as the change is self-explanatory.